Feature/player/input action

+++ Source/OriginalSinPrj/Public/Player/Struct/AbilityDataBuffer.h
@@ -47,13 +47,13 @@ public:
 	float AddedDamage = 0.0f;
 
 	UPROPERTY()
-	int32 MaxMana = 5;
+	float MaxMana = 5;
 
 	UPROPERTY()
-	int32 CurrentMana = 5;
+	float CurrentMana = 5;
 
 	UPROPERTY()
-	int32 NeedMana = 0;
+	float NeedMana = 0;
 
 	UPROPERTY()
 	TObjectPtr<AActor> DamageCauser = nullptr;
